refactor(web): split chat center modules
